chore: maintenance cleanup — zero warnings, dead code removed, systems wired
Clippy: 41 warnings → 0 warnings
- All is_multiple_of, collapsible ifs, format! cleanup applied
- Blanket #![allow(dead_code)] replaced with targeted #[allow] on specific items
- Dead methods removed: GameState::check_achievements(), update_highest_sector()
- Legacy prestige fields zeroed in complete_voyage()
- Consolidated sector_faction naming
- All Navigator match arms verified

Wiring fixes:
- Faction reputation affects enemy stats in battle (+20% hostile, -15% allied)
- Battle kills emit FactionRepChange events (-15 per ship)
- Raid completion emits FactionRepChange (-25 for planet faction)
- All GameEvent variants have handlers in process_events
- Trade/mission events fully processed

25,544 lines. 237 tests. 0 clippy warnings.

+++ src/main.rs
@@ -1,4 +1,4 @@
-#![allow(dead_code)] // Game systems defined for future wiring — intentional
+ // Game systems defined for future wiring — intentional
 
 mod engine;
 mod rendering;
@@ -365,7 +365,7 @@ fn main() -> io::Result<()> {
                             state.sector,
                             &sector_faction,
                             &state.cargo,
-                            state.sector as u32 + state.total_battles as u32,
+                            state.sector + state.total_battles as u32,
                         );
                         for result in contraband_results {
                             event_bus.emit(engine::events::GameEvent::ContrabandDetected {
